% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_levels_vec.R
\name{get_levels_vec}
\alias{get_levels_vec}
\title{Generate Opinion Levels}
\usage{
get_levels_vec(num_opinions)
}
\arguments{
\item{num_opinions}{Number of opinions}
}
\value{
A numeric vector of length \code{num_opinions} around 0
}
\description{
Creates a vector of discrete opinion levels based on the specified number
of opinions. The levels are evenly spaced around 0
}
\details{
The levels are generated using an evenly spaced sequence
}
\examples{
# Generate 3 opinion levels
get_levels_vec(3)
# Returns: -2 -1  1  2

}
\author{
Victor Bernal
}
