% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_palette.R
\name{get_palette}
\alias{get_palette}
\title{get_palette}
\usage{
get_palette(num_opinions)
}
\arguments{
\item{num_opinions}{number of opinions}
}
\value{
palette \code{x}.
}
\description{
get_palette
}
\examples{
#---------------------------------
# Initialize color palette
# get_palette (num_opinions)
#---------------------------------
}
\author{
Victor Bernal
}
