% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visual_histo.R
\name{visual_histo}
\alias{visual_histo}
\title{visual_histo}
\usage{
visual_histo(opinion_history, num_opinions)
}
\arguments{
\item{opinion_history}{opinion_history}

\item{num_opinions}{num_opinions}
}
\description{
visual_histo
}
\examples{
#---------------------------------
# visual_histo
# visual_histo(opinion_history, num_opinions)
#---------------------------------
}
\author{
Victor Bernal \email{victor.arturo.bernal@gmail.com}
}
